+++ packages/analyzer/src/javascript/resolve-specifier-node.ts
@@ -67,25 +67,23 @@ export const resolve =
             },
       });
 
-      let relativeSpecifierUrl =
-          relative(dirname(documentPath), resolvedSpecifier) as FileRelativeUrl;
-
-      // If this is a component-style URL layout, rewrite the path to reach out
-      // of the current package, rather than directly into the component
-      // directory.
+      let effectiveDocumentPath = documentPath;
       if (componentInfo !== undefined) {
         const {packageName, rootDir, componentDir} = componentInfo;
-        if (pathIsInside(resolvedSpecifier, componentDir)) {
-          const componentDirRelativeToPackage =
-              relative(join(rootDir, packageName), rootDir);
-          const relativePathFromComponentDir =
-              relative(componentDir, resolvedSpecifier);
-          relativeSpecifierUrl =
-              join(
-                  componentDirRelativeToPackage,
-                  relativePathFromComponentDir) as FileRelativeUrl;
+        const isRootPackageRequest = !pathIsInside(documentPath, componentDir);
+        if (isRootPackageRequest) {
+          // Special handling for servers like Polyserve which, when serving a
+          // package "foo", will map the URL "/components/foo" to the root
+          // package directory, so that "foo" can make correct relative path
+          // references to its dependencies.
+          const rootRelativePath = relative(rootDir, documentPath);
+          effectiveDocumentPath =
+              join(componentDir, packageName, rootRelativePath);
         }
       }
+
+      let relativeSpecifierUrl = relative(
+          dirname(effectiveDocumentPath), resolvedSpecifier) as FileRelativeUrl;
 
       if (isWindows()) {
         // normalize path separators to URL format
@@ -98,5 +96,4 @@ export const resolve =
       }
 
       return relativeSpecifierUrl;
-
     };