+++ packages/analyzer/CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-<!-- ## Unreleased -->
+## Unreleased
+* Fix module resolution in the case of root packages served from a component
+  directory.
 <!-- Add new, unreleased changes here. -->
 
 ## [3.0.0-pre.21] - 2018-04-05